src/sprucfluo/requirements.txt
webdataset
datasets
transformers
torchdata
fsspec
zstandard